Refactor reusable components out of Uptime Details
